+++ CHANGELOG.md
@@ -1,13 +1,13 @@
 #### CHANGELOG:
-- 2016-01-03, Version 1.4.3
+- 2016-01-20, Version 1.4.3
   * More detailed version for version feature.
   * Dynamic set permissions.
   * Modify settin bar icon.
   * Add background position item, intelligent adjusting background position.
 
 - 2015-12-25, Version 1.4.2
-  * Add new background: [NASA Image Galleries](http://www.nasa.gov/multimedia/imagegallery/index.html)
-  * Add special day to Setting bar.
+  * Add new background: [NASA Astronomy Picture of the Day](http://apod.nasa.gov/apod/astropix.html)
+  * Add special day(SimpTab Images) to Setting bar.
 
 - 2015-12-23, Version 1.4.1
   * Add 'senior' and 'normal' topsites.( old topsites only 'simple' mode.)